backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
yfinance>=0.2.65
exchange-calendars>=4.13.0
holidays>=0.91
pandas>=2.2.0
numpy>=2.1.0
SQLAlchemy>=2.0.40
psycopg[binary]>=3.2.9
pydantic>=2.11.0
python-dotenv>=1.1.1
python-multipart>=0.0.20
openai>=1.99.0
zhipuai>=2.1.5
sniffio>=1.3.1
pytest>=8.4.0
